requirements.txt
llama-index
llama-index-embeddings-huggingface
llama-index-llms-ollama
chromadb
sentence-transformers
pypdf
python-docx
pandas
numpy
openpyxl
python-pptx